Refactor TACE helpers into shared e3nn_nn and blocks packages.
Move IrrepsLinear/paths/scatter-TP/gated nonlinear into e3nn_nn and reuse
RadialMLP for channel MLPs so tace/ keeps only core, interaction, and
Cartesian-specific code.
