+++ docs/developer_guide.rst
@@ -438,7 +438,7 @@ DPA4 (Li et al., 2026, arXiv:2606.02419) lives under :code:`enerzyme/models/dpa4
 TACE (:code:`tace`)
 ^^^^^^^^^^^^^^^^^^^
 
-Xu et al. (arXiv:2509.14961; Cartesian-3j arXiv:2512.16882) lives under :code:`enerzyme/models/tace/`, adapted from `xvzemin/tace <https://github.com/xvzemin/tace>`_ (MIT). Register via :code:`get_ff_core("tace")`. Core param :code:`tensor_basis` selects spherical e3nn CGTP (:code:`CgtpInteraction` + :code:`CgtpACE`) or Cartesian ICT (:code:`cartnn` vendored from tace v0.1.0 + :code:`cartesian/core_blocks.py`). Scope includes edge embedding/update, BB element-aware residual, and density/avg scatter-norm; TECE / SO2 / RRA / ZBL / LES / UIE stay out. Emit :code:`atom_feature` for :code:`SimpleReadout`. Examples: :code:`tace_layers_example.yaml`, :code:`tace_cartesian_layers_example.yaml`. Tests: :code:`test/test_tace_*.py`.
+Xu et al. (arXiv:2509.14961; Cartesian-3j arXiv:2512.16882) lives under :code:`enerzyme/models/tace/` (:code:`core.py`, :code:`interaction.py`), adapted from `xvzemin/tace <https://github.com/xvzemin/tace>`_ (MIT). Shared flat-Irreps helpers (:code:`IrrepsLinear`, :code:`generate_paths`, :code:`O3ScatterTensorProduct`, :code:`get_gated_nonlinear`, …) live in :code:`enerzyme/models/e3nn_nn/`; radial channel MLPs use :code:`blocks.radial_mlp.RadialMLP`. Register via :code:`get_ff_core("tace")`. Core param :code:`tensor_basis` selects spherical e3nn CGTP or Cartesian ICT (:code:`cartnn` vendored from tace v0.1.0 + :code:`cartesian/`). Scope includes edge embedding/update, BB element-aware residual, and density/avg scatter-norm; TECE / SO2 / RRA / ZBL / LES / UIE stay out. Emit :code:`atom_feature` for :code:`SimpleReadout`. Examples: :code:`tace_layers_example.yaml`, :code:`tace_cartesian_layers_example.yaml`. Tests: :code:`test/test_tace_*.py`.
 
 So3krates (:code:`so3krates`)
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^
