+++ docs/user_guide/models/architecture_catalog.rst
@@ -83,7 +83,7 @@ External models are declared under :code:`Modelhub.external_FFs` with the same :
 
 **DPA4** (:code:`architecture: dpa4`) implements the EMFA SO(2) descriptor of Li et al. (2026, arXiv:2606.02419) under :code:`enerzyme/models/dpa4/` (:code:`core.py`, :code:`interaction.py`, :code:`so2.py`). Shared SO(3) helpers (C³ envelope, Apache e3x Lebedev tables / :code:`S2LebedevProjector`, packed indexing, :code:`FocusSO2Linear`, focus gated activation, degree RMSNorm, Bessel×C³ RBF, quaternion edge frames + shared e3nn/:code:`Jd` Wigner-D via :code:`A R Aᵀ` into DPA4's Cartesian basis; flat :code:`SO3Grid` / Lebedev behind :code:`S2GridProjector`) live in :code:`enerzyme/models/so3/`. The Core owns the radial cache, geometric/environment seed, m-truncated interactions, and Lebedev-grid equivariant FFN. It emits :code:`atom_feature` / :code:`atom_sphere_feature`, so the standard :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force` stack applies. Example: :code:`enerzyme/config/dpa4_layers_example.yaml`. Tests: :code:`test/test_dpa4_core.py`, :code:`test/test_dpa4_parity_ops.py`. Enerzymette only needs :code:`architecture: dpa4` plus a resolved :code:`config.yaml`.
 
-**TACE** (:code:`architecture: tace`) is a native port of Xu et al. (arXiv:2509.14961; Cartesian-3j arXiv:2512.16882) under :code:`enerzyme/models/tace/`, adapted from `xvzemin/tace <https://github.com/xvzemin/tace>`_ (MIT). The Core is the ACE message/product stack only (edge update + CgtpInteraction / Cartesian ICT contraction + CgtpACE / Cartesian product); embeddings, property heads, ZBL/LES/UIE stay outside. Switch backends with Core param :code:`tensor_basis: spherical|cartesian` (default spherical). Spherical uses e3nn CGTP paths (Lmax/lmax split, BB element-aware residual, optional density scatter-norm). Cartesian vendors :code:`cartnn` from tace v0.1.0. TECE / SO2 / RRA are not included. Emits :code:`atom_feature` with :code:`feature_irreps` / :code:`dim_feature_out` for :code:`SimpleReadout`. Examples: :code:`enerzyme/config/tace_layers_example.yaml`, :code:`tace_cartesian_layers_example.yaml`. Tests: :code:`test/test_tace_core.py`, :code:`test/test_tace_spherical_ops.py`, :code:`test/test_tace_cartesian_ops.py`. Enerzymette only needs :code:`architecture: tace` plus a resolved :code:`config.yaml`.
+**TACE** (:code:`architecture: tace`) is a native port of Xu et al. (arXiv:2509.14961; Cartesian-3j arXiv:2512.16882) under :code:`enerzyme/models/tace/` (:code:`core.py`, :code:`interaction.py`), adapted from `xvzemin/tace <https://github.com/xvzemin/tace>`_ (MIT). Shared flat-Irreps helpers live in :code:`enerzyme/models/e3nn_nn/`; radial MLPs use :code:`blocks.radial_mlp.RadialMLP`. The Core is the ACE message/product stack only (edge update + CgtpInteraction / Cartesian ICT + CgtpACE); embeddings, property heads, ZBL/LES/UIE stay outside. Switch backends with Core param :code:`tensor_basis: spherical|cartesian` (default spherical). TECE / SO2 / RRA are not included. Emits :code:`atom_feature` with :code:`feature_irreps` / :code:`dim_feature_out` for :code:`SimpleReadout`. Examples: :code:`enerzyme/config/tace_layers_example.yaml`, :code:`tace_cartesian_layers_example.yaml`. Tests: :code:`test/test_tace_core.py`, :code:`test/test_tace_spherical_ops.py`, :code:`test/test_tace_cartesian_ops.py`. Enerzymette only needs :code:`architecture: tace` plus a resolved :code:`config.yaml`.
 
 **So3krates** (:code:`architecture: so3krates`) is a native port of Frank et al. (NeurIPS 2022) under :code:`enerzyme/models/so3krates/`, following the So3krates-torch EuclideanTransformer (fused FeatureBlock + GeometricBlock + InteractionBlock). Shared :code:`RealSphericalHarmonics` and :code:`L0Contraction` live in :code:`enerzyme/models/so3/`. The Core emits :code:`atom_feature` (invariant stream ``x``, :code:`feature_irreps: "Fx0e"`) and :code:`atom_sphere_feature` (SPHC ``χ`` with shape ``[N, m_tot]``, **not** eSCN/EquiformerV2's ``[N, (lmax+1)^2, C]`` — :code:`SphereSampleReadout` does not apply). Default stacks use :code:`BernsteinRBF` + :code:`SimpleReadout` + :code:`EnergyReduce` + :code:`Force`. Long-range physics (ZBL / electrostatics / dispersion) belong in post-core layers, not inside the Core. Example: :code:`enerzyme/config/so3krates_layers_example.yaml`. Parity: :code:`test/test_so3krates_parity_ops.py`.
