+++ enerzyme/models/e3nn_nn/tools.py
@@ -95,6 +95,90 @@ def tp_out_irreps_with_instructions(
     return irreps_out, instructions
 
 
+def satisfy(l1: int, l2: int, restriction: Optional[str] = None) -> bool:
+    """Angular-momentum restriction used by ACE / TACE path filters."""
+    if restriction is None:
+        return True
+    if restriction == "<":
+        return l1 < l2
+    if restriction == "<=":
+        return l1 <= l2
+    if restriction == ">":
+        return l1 > l2
+    if restriction == ">=":
+        return l1 >= l2
+    if restriction == "==":
+        return l1 == l2
+    if restriction == "!=":
+        return l1 != l2
+    raise ValueError(f"Unknown restriction: {restriction}")
+
+
+def generate_paths(
+    irreps_out: Irreps,
+    irreps_in1: Irreps,
+    irreps_in2: Irreps,
+    *,
+    l1l2: Optional[str] = None,
+    l2l3: Optional[str] = None,
+    l3l1: Optional[str] = None,
+    e3nn_mode: str = "uvu",
+    trainable: bool = False,
+    identical_inputs: bool = False,
+) -> Tuple[List[Tuple[int, int, int, str, bool]], Irreps]:
+    """Build e3nn TensorProduct instructions with optional l-restrictions.
+
+    Adapted from xvzemin/tace (MIT).
+    """
+    irreps_out = Irreps(irreps_out)
+    irreps_in1 = Irreps(irreps_in1)
+    irreps_in2 = Irreps(irreps_in2)
+    if identical_inputs and irreps_in1 != irreps_in2:
+        raise ValueError("identical_inputs requires matching input irreps")
+
+    e3nn_paths: List[Tuple[int, int, int, str, bool]] = []
+    e3nn_out_irreps: List[Tuple[int, o3.Irrep]] = []
+
+    for _, ir_out in irreps_out:
+        for i, (mul, ir1) in enumerate(irreps_in1):
+            for j, (_, ir2) in enumerate(irreps_in2):
+                l1, l2, l3 = ir1.l, ir2.l, ir_out.l
+                if (
+                    ir_out in ir1 * ir2
+                    and satisfy(l1, l2, l1l2)
+                    and satisfy(l2, l3, l2l3)
+                    and satisfy(l3, l1, l3l1)
+                ):
+                    if identical_inputs and i == j and (l1 + l2 - l3) % 2 == 1:
+                        continue
+                    k = len(e3nn_out_irreps)
+                    e3nn_out_irreps.append((mul, (ir_out.l, ir_out.p)))
+                    e3nn_paths.append(
+                        (i, j, k, e3nn_mode, e3nn_mode == "uvu" or trainable)
+                    )
+    return e3nn_paths, Irreps(e3nn_out_irreps)
+
+
+def to_possible_tp_irreps(
+    irreps_in1: Irreps,
+    irreps_in2: Irreps,
+    parity: bool,
+    lmax: Optional[int] = None,
+) -> Irreps:
+    """Irreps closed under TP(in1, in2), truncated to ``lmax`` / natural parity."""
+    irreps_in1 = Irreps(irreps_in1)
+    irreps_in2 = Irreps(irreps_in2)
+    lmax = irreps_in2.lmax if lmax is None else lmax
+    irrep_set = {
+        ir3
+        for _, ir1 in irreps_in1
+        for _, ir2 in irreps_in2
+        for ir3 in ir1 * ir2
+        if ir3.l <= lmax and (parity or ir3.p == (-1) ** ir3.l)
+    }
+    return Irreps(sorted(irrep_set, key=lambda ir: (ir.l, ir.p))).regroup()
+
+
 @compile_mode("script")
 class reshape_irreps(torch.nn.Module):
     def __init__(self, irreps: Irreps) -> None:
