--- enerzyme/models/tace/spherical/__init__.py
@@ -1,6 +1,0 @@
-"""Spherical TACE blocks."""
-
-from .interaction import CgtpInteraction
-from .product import CgtpACE
-
-__all__ = ["CgtpInteraction", "CgtpACE"]
